Add in git info
We wanted to pull in the git information into the /info endpoint of the
actuator. With this we can see the commit id and branch. We may be able
to pull in the build number, which would also be nice, though we might
need to see how that works out.

That best would be to bump the artifact version on each build.